code/pyproject.toml
[project]
name = "chainlit-sandbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.6.3",
    "openai>=2.24.0",
    "python-dotenv>=1.2.2",
]
